Added offset to polyline.
