+++ src/GShark.Test.XUnit/Geometry/PolygonTests.cs
@@ -2,6 +2,7 @@
 using GShark.Core;
 using GShark.Geometry;
 using System;
+using System.Linq;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -138,19 +139,40 @@ namespace GShark.Test.XUnit.Geometry
         public void It_Returns_A_Polygon_Transformed_In_NurbsCurve()
         {
             // Arrange
-            Polyline poly = new Polygon(Planar2D);
-            NurbsCurve polyCurve = poly.ToNurbs();
+            Polyline polygon = new Polygon(Planar2D);
             double lengthSum = 0.0;
 
-            // Assert
-            polyCurve.Degree.Should().Be(1);
-            // ToDo: has to be completed refactoring the polyline.
-            for (int i = 1; i < poly.SegmentsCount; i++)
-            {
-                lengthSum += poly.Segments[i].Length;
+            // Act
+            NurbsCurve polygonCurve = polygon.ToNurbs();
 
-                // Planar2D[i - 1].Equals(poly.PointAtLength(lengthSum)).Should().BeTrue();
+            // Assert
+            polygonCurve.Degree.Should().Be(1);
+            polygonCurve.ControlPointLocations[0]
+                .EpsilonEquals(polygonCurve.ControlPointLocations.Last(), GSharkMath.MinTolerance).Should().BeTrue();
+            for (int i = 0; i < polygon.SegmentsCount; i++)
+            {
+                lengthSum += polygon.Segments[i].Length;
+                polygon[i + 1].EpsilonEquals(polygonCurve.PointAtLength(lengthSum), GSharkMath.MaxTolerance).Should().BeTrue();
             }
+        }
+
+        [Theory]
+        [InlineData(0, 7.071068)]
+        [InlineData(1, 12.205361)]
+        [InlineData(2, 5.481013)]
+        [InlineData(3, 7.071068)]
+        [InlineData(4, 7.071068)]
+        public void Returns_The_Offset_Of_A_Polygon(int vertex, double expectedDistance)
+        {
+            // Arrange
+            Polygon polygon = new Polygon(PolygonTests.Planar2D);
+            double offset = 5;
+
+            Polyline offsetPolygon = polygon.Offset(offset, Plane.PlaneXY);
+
+            // Assert
+            polygon[vertex].DistanceTo(offsetPolygon[vertex]).Should()
+                .BeApproximately(expectedDistance, GSharkMath.MaxTolerance);
         }
 
         [Fact]
