+++ src/GShark/Geometry/Polyline.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using GShark.Interfaces;
+using GShark.Operation;
 
 namespace GShark.Geometry
 {
@@ -38,11 +39,6 @@ namespace GShark.Geometry
         /// A polyline is considered closed, if its start and end point are identical.
         /// </summary>
         public bool IsClosed => this[0] == this[Count - 1];
-
-        /// <summary>
-        /// Gets the domain of the polyline.
-        /// </summary>
-        public Interval Domain => new Interval(0, SegmentsCount);
 
         /// <summary>
         /// Gets the starting point of the polyline.
@@ -387,11 +383,70 @@ namespace GShark.Geometry
         }
 
         /// <summary>
+        /// Computes the offset of the polyline.
+        /// </summary>
+        /// <param name="distance">The distance of the offset.</param>
+        /// <param name="pln">The plane for the offset operation.</param>
+        /// <returns>The offset polyline.</returns>
+        public Polyline Offset(double distance, Plane pln)
+        {
+            if (distance == 0.0)
+            {
+                return this;
+            }
+
+            int iteration = (IsClosed) ? Count : Count - 1;
+
+            Point3[] offsetPts = new Point3[Count];
+            List<Line> segments = Segments;
+            Line[] offsetSegments = new Line[segments.Count + 1];
+
+            for (int i = 0; i < iteration; i++)
+            {
+                int k = (i == iteration - 1 && IsClosed) ? 0 : i;
+                if (i == iteration - 1 && k == 0)
+                {
+                    goto Intersection;
+                }
+
+                Vector3 vecOffset = Vector3.CrossProduct(segments[k].Direction, pln.ZAxis).Amplify(distance);
+                Transform xForm = Core.Transform.Translation(vecOffset);
+                offsetSegments[k] = segments[k].Transform(xForm);
+
+                if (i == 0 && IsClosed)
+                {
+                    continue;
+                }
+                if (k == 0 && !IsClosed)
+                {
+                    offsetPts[k] = offsetSegments[k].StartPoint;
+                    continue;
+                }
+
+                Intersection:
+                bool ccx = Intersect.LineLine(offsetSegments[(i == iteration - 1 && IsClosed) ? iteration - 2 : k - 1], offsetSegments[k], out Point3 pt, out _, out _, out _);
+                if (!ccx)
+                {
+                    continue;
+                }
+
+                offsetPts[k] = pt;
+
+                if (i == iteration - 1)
+                {
+                    offsetPts[(IsClosed) ? i : i + 1] = (IsClosed) ? offsetPts[0] : offsetSegments[k].EndPoint;
+                }
+            }
+
+            return new Polyline(offsetPts);
+        }
+
+        /// <summary>
         /// Compute the segments length and removes the segments which are shorter than a tolerance.
         /// </summary>
         /// <param name="vertices">Points used to create the polyline.</param>
         /// <returns>A cleaned collections of points if necessary otherwise the same collection of points.</returns>
-        protected static IList<Point3> CleanVerticesForShortLength(IList<Point3> vertices)
+        private static IList<Point3> CleanVerticesForShortLength(IList<Point3> vertices)
         {
             int[] coincidenceFlag = new int[vertices.Count];
             coincidenceFlag[0] = 0;
